[mlir][tosa]: Add Tenosr Encoding Lowering in tosa-to-arith
Add support for lowering tensors with the `encoding` attribute in the
`tosa::ConstOp` rewrite pattern.

Because `arith::ConstantOp` inherits from the `AllTypesMatch` interface
it requires its values and results types match. This means we need to
materialize a `tensor::CastOp` in the case that the there is a type
mismatch between values and results in the `tosa::ConstOp` e.g. when the
values don't have an encoding attribute in their type but the results
do.

Signed-off-by: Jack Frankland <jack.frankland@arm.com>